Remove logo from nav bar
Thought it would be easy and straightforward to add the image to the
homepage and nav bar, but it seems that working with images is a bit
more complex in css then I thought. I'll hold off on adding it to the
page until I've learned how to deal with images specifically.